git console test

+++ src/br/com/monster/portal/adm/controller/CategoriaController.java
@@ -1,5 +1,7 @@
 package br.com.monster.portal.adm.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.com.monster.portal.model.Categoria;
 import br.com.monster.portal.modelDao.CategoriaDao;
+import br.com.monster.portal.modelDao.RelatorioDao;
+import br.com.monster.portal.security.EnumEntidade;
+import br.com.monster.portal.security.EnumMetodo;
+import br.com.monster.portal.security.Permissoes;
 
 /*
  * @author Filipe A. Pimenta
@@ -21,10 +27,18 @@ import br.com.monster.portal.modelDao.CategoriaDao;
 @Transactional
 @Controller
 public class CategoriaController {
-	
 
-		@Autowired
-		CategoriaDao dao;
+	@Autowired
+	CategoriaDao dao;
+		
+	/*
+	 * Registros e permissões
+	 */
+	// Define Entidade
+	private EnumEntidade entidade = EnumEntidade.CATEGORIA;
+	// Consulta a interface RelatorioDao
+	@Autowired
+	RelatorioDao relatorio;
 		
 		/*
 
@@ -42,17 +56,20 @@ public class CategoriaController {
 		 *	@param Categoria - O Objeto principal para a criação
 		 *	@return String - Manipulado pelo Spring para o método read (leitura)
 		 */
-		
-		@RequestMapping("Admin/CreateCategoria")
-		public String create(@Valid Categoria categoria, BindingResult result) {			
-			if(result.hasErrors()) {														//	Se houver erro na validação
-			    return "forward:categoria";													//	Volte para a página categoria
+	
+	@RequestMapping("Admin/CreateCategoria")
+	public String create(HttpSession session, @Valid Categoria categoria, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.CRIAR, entidade)){				//	Checar Permissão
+			if(result.hasErrors()) {											//	Se houver erro na validação
+			    return "forward:categoria";										//	Volte para a página de adição
 			} else {
-				dao.create(categoria);														//	Ação no banco
-				return "redirect:categoria";												//	Retorna para o método Read
+				dao.create(categoria);											//	Ação no banco
+				relatorio.gerarRelatorio(session, EnumMetodo.CRIAR, entidade);	//	Relatório
+				return "redirect:categoria";									//	Retorna para o método Read
 			}
-			
 		}
+		return "403";
+	}
 
 		/*
 		 * -------------------------
@@ -65,12 +82,14 @@ public class CategoriaController {
 		 *	@return String - Página read (leitura)
 		 */
 		
-		@RequestMapping("Admin/categoria")
-		public String Read(Model model) {
-			model.addAttribute("categorias", dao.read());									//	Consulta o Banco e coloca na variável da página
-			return "admin/Categoria/read";												//	Retorna para á página JSP
+	@RequestMapping("Admin/categoria")
+	public String read(HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.LER, entidade)){				//	Consulta Permissão
+			model.addAttribute("categorias", dao.read());						//	Consulta o Banco e coloca na variável da página
+			return "admin/Categoria/read";										//	Retorna para á página JSP
 		}
-
+		return "403";
+	}
 		/*
 		 * -------------------------
 		 * 			Update			
@@ -84,15 +103,20 @@ public class CategoriaController {
 		 *	@return String - Manipulado pelo Spring para o método read (leitura)
 		 */
 
-		@RequestMapping("Admin/UpdateCategoria")
-		public String update(@Valid Categoria categoria, BindingResult result) {
-			if(result.hasErrors()) {														//	Se houver erro na validação
-			    return "forward:categoria";													//	Volte
+	@RequestMapping("Admin/UpdateCategoria")
+	public String update(HttpSession session, @Valid Categoria categoria, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){			//	Consulta Permissão
+			if(result.hasErrors()) {											//	Se houver erro na validação
+			    return "forward:categoria";										//	Volte
 			} else {
-				dao.update(categoria);														//	Ação no banco
-				return "redirect:categoria";												//	Retorna para o método Read
+				dao.update(categoria);											//	Ação no banco
+				relatorio.gerarRelatorio(session, EnumMetodo.ATUALIZAR, entidade);	//	Gera Relatório e armazena no banco
+				return "redirect:categoria";									//	Retorna para o método Read
 			}
 		}
+		return "403";
+		
+	}
 
 		/*
 		 * -------------------------
@@ -107,11 +131,15 @@ public class CategoriaController {
 		 *	@return void - deletar não precisa de um retorno
 		 */
 		
-		@RequestMapping("Admin/DeleteCategoria")
-		public void delete(Long id) {
-		  dao.delete(id);																	//	Ação no banco
+	@RequestMapping("Admin/DeleteCategoria")
+	public void delete(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.EXCLUIR, entidade)){			//	Consulta a permissão
+			dao.delete(id);														//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.EXCLUIR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
 		}
-		
+	} 
+	
 	
 		/*
 		 * -------------------------
@@ -126,10 +154,14 @@ public class CategoriaController {
 		 *	@return void - restaurar não precisa de um retorno
 		 */
 		
-		@RequestMapping("Admin/RestoreCategoria")
-		public void restore(Long id) {
-			  dao.restore(id);																//	Ação no banco
+	@RequestMapping("Admin/RestoreCategoria")
+	public void restore(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.RESTAURAR, entidade)){			//	Consulta a permissão
+			dao.restore(id);													//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.RESTAURAR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
 		}
+	}
 
 		/*
 		 * -------------------------
@@ -144,10 +176,14 @@ public class CategoriaController {
 		 *	@return String - retorna uma página JSP
 		 */
 		
-		@RequestMapping("Admin/FindCategoria")
-		public String find_one(Long id, Model model) {
-		  model.addAttribute("categorias", dao.findOne(id));								//	Consulta o Banco e coloca na variável da página
-		  return "admin/Categoria/edt";														//	Retorna para a página JSP edt
-		}
-		
+	@RequestMapping("Admin/FindCategoria")
+	public String Find(Long id, HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){			//	Consulta a permissão
+			model.addAttribute("categoria", dao.findOne(id));					//	Consulta o Banco e coloca na variável da página
+			return "admin/Categoria/edt";										//	Consulta o Banco e coloca na variável da página
+		}	
+		return"403";
+	}
+	
 }
+
