+++ src/br/com/monster/portal/adm/controller/FuncionarioController.java
@@ -1,5 +1,7 @@
 package br.com.monster.portal.adm.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import br.com.monster.portal.model.Funcionario;
 import br.com.monster.portal.modelDao.CargoDao;
 import br.com.monster.portal.modelDao.FuncionarioDao;
+import br.com.monster.portal.modelDao.RelatorioDao;
+import br.com.monster.portal.security.EnumEntidade;
+import br.com.monster.portal.security.EnumMetodo;
+import br.com.monster.portal.security.Permissoes;
 
 /*
  * @author Filipe A. Pimenta
@@ -28,6 +34,15 @@ public class FuncionarioController {
 	
 	@Autowired
 	CargoDao cargo_dao;
+	
+	/*
+	 * Registros e permissões
+	 */
+	// Define Entidade
+	private EnumEntidade entidade = EnumEntidade.FUNCIONARIO;
+	// Consulta a interface RelatorioDao
+	@Autowired
+	RelatorioDao relatorio;
 	
 	/*
 
@@ -47,15 +62,17 @@ public class FuncionarioController {
 	 */
 	
 	@RequestMapping("Admin/CreateFuncionario")
-	public String create(@Valid Funcionario funcionario, BindingResult result) {			
-		
-		if(result.hasErrors()) {															//	Se houver erro na validação
-		    return "forward:funcionario";													//	Volte para a página funcionario
-		} else {
-			dao.create(funcionario);														//	Ação no banco
-			return "redirect:funcionario";													//	Retorna para o método Read
+	public String create(HttpSession session, @Valid Funcionario funcionario, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.CRIAR, entidade)){				//	Checar Permissão
+			if(result.hasErrors()) {											//	Se houver erro na validação
+			    return "forward:adicionar_funcionario";								//	Volte para a página de adição
+			} else {
+				dao.create(funcionario);						//	Ação no banco
+				relatorio.gerarRelatorio(session, EnumMetodo.CRIAR, entidade);	//	Relatório
+				return "redirect:funcionario";										//	Retorna para o método Read
+			}
 		}
-		
+		return "403";
 	}
 
 	/*
@@ -70,9 +87,12 @@ public class FuncionarioController {
 	 */
 	
 	@RequestMapping("Admin/funcionario")
-	public String Read(Model model, Funcionario funcionario) {
-		model.addAttribute("funcionarios", dao.read());										//	Consulta o Banco e coloca na variável da página
-		return "admin/Funcionario/read";													//	Retorna para á página JSP
+	public String read(HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.LER, entidade)){					//	Consulta Permissão
+			model.addAttribute("funcionarios", dao.read());							//	Consulta o Banco e coloca na variável da página
+			return "admin/Funcionario/read";										//	Retorna para á página JSP
+		}
+		return "403";
 	}
 
 	/*
@@ -89,13 +109,17 @@ public class FuncionarioController {
 	 */
 
 	@RequestMapping("Admin/UpdateFuncionario")
-	public String update(@Valid Funcionario funcionario, BindingResult result) {
-		if(result.hasErrors()) {															//	Se houver erro na validação
-		    return "forward:funcionario";													//	Volte
-		} else {
-			dao.update(funcionario);														//	Ação no banco
-			return "redirect:funcionario";													//	Retorna para o método Read
+	public String update(HttpSession session, @Valid Funcionario funcionario, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){				//	Consulta Permissão
+			if(result.hasErrors()) {												//	Se houver erro na validação
+			    return "forward:funcionario";										//	Volte
+			} else {
+				dao.update(funcionario);											//	Ação no banco
+				relatorio.gerarRelatorio(session, EnumMetodo.ATUALIZAR, entidade);	//	Gera Relatório e armazena no banco
+				return "redirect:funcionario";										//	Retorna para o método Read
+			}
 		}
+		return "403";
 	}
 
 	/*
@@ -112,9 +136,13 @@ public class FuncionarioController {
 	 */
 	
 	@RequestMapping("Admin/DeleteFuncionario")
-	public void delete(Long id) {
-	  dao.delete(id);																		//	Ação no banco
-	}
+	public void delete(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.EXCLUIR, entidade)){			//	Consulta a permissão
+			dao.delete(id);														//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.EXCLUIR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
+		}
+	} 
 	
 	
 	/*
@@ -131,8 +159,12 @@ public class FuncionarioController {
 	 */
 	
 	@RequestMapping("Admin/RestoreFuncionario")
-	public void restore(Long id) {
-		  dao.restore(id);																	//	Ação no banco
+	public void restore(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.RESTAURAR, entidade)){			//	Consulta a permissão
+			dao.restore(id);													//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.RESTAURAR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
+		}
 	}
 
 	/*
@@ -149,9 +181,12 @@ public class FuncionarioController {
 	 */
 	
 	@RequestMapping("Admin/FindFuncionario")
-	public String Find(Model model, Long id) {
-		model.addAttribute("funcionarios", dao.findOne(id));								//	Consulta o Banco e coloca na variável da página
-		return "admin/Funcionario/edt";														//	Retorna para a página JSP edt
+	public String Find(Long id, HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){			//	Consulta a permissão
+			model.addAttribute("funcionario", dao.findOne(id));					//	Consulta o Banco e coloca na variável da página
+			return "admin/Funcionario/edt";										//	Consulta o Banco e coloca na variável da página
+		}
+		return"403";
 	}
-
+	
 }
