+++ src/br/com/monster/portal/adm/controller/PedidoController.java
@@ -1,5 +1,7 @@
 package br.com.monster.portal.adm.controller;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,12 +11,16 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import br.com.monster.portal.carrinho.Carrinho;
 import br.com.monster.portal.model.Cliente;
 import br.com.monster.portal.model.Pedido;
 import br.com.monster.portal.modelDao.ClienteDao;
 import br.com.monster.portal.modelDao.PedidoDao;
 import br.com.monster.portal.modelDao.ProdutoDao;
-import br.com.monster.portal.carrinho.Carrinho;
+import br.com.monster.portal.modelDao.RelatorioDao;
+import br.com.monster.portal.security.EnumEntidade;
+import br.com.monster.portal.security.EnumMetodo;
+import br.com.monster.portal.security.Permissoes;
 
 /*
  * @author Filipe A. Pimenta
@@ -36,6 +42,15 @@ public class PedidoController {
 	ClienteDao dao_cli;
 	
 	/*
+	 * Registros e permissões
+	 */
+	// Define Entidade
+	private EnumEntidade entidade = EnumEntidade.PEDIDO;
+	// Consulta a interface RelatorioDao
+	@Autowired
+	RelatorioDao relatorio;
+	
+	/*
 
 	 |==================================|
 	 |				Métodos				|
@@ -53,23 +68,27 @@ public class PedidoController {
 	 */
 	
 	@RequestMapping("Admin/adicionar_pedido")
-	public String create_page(Model model, Pedido pedido) {
-		model.addAttribute("pedidos", dao.read());
-		model.addAttribute("produtos", dao_prod.read());
-		model.addAttribute("clientes", dao_cli.read());
-		return "admin/Pedido/adicionar";
+	public String create_page(HttpSession session, Model model, Pedido pedido) {
+		if(Permissoes.checar(session, EnumMetodo.CRIAR, entidade)){			//	Consulta a permissão
+			model.addAttribute("pedidos", dao.read());
+			model.addAttribute("produtos", dao_prod.read());
+			model.addAttribute("clientes", dao_cli.read());
+			return "admin/Pedido/adicionar";
+		}
+		return "403";
 	}
 	
 	@RequestMapping("Admin/CreatePedido")
-	public String create(@Valid Pedido pedido, Carrinho carrinho, Cliente cliente, BindingResult result) {
-
-		if(result.hasErrors()) {											//	Se houver erro na validação
-		    return "forward:adicionar_pedido";								//	Volte para a página de adição
-		} else {
-			dao.create(pedido, carrinho, cliente);									//	Ação no banco
-			return "redirect:adicionar_pedido";								//	Retorna para o método Read
+	public String create(HttpSession session, @Valid Pedido pedido, Carrinho carrinho, Cliente cliente, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.CRIAR, entidade)){			//	Consulta a permissão
+			if(result.hasErrors()) {											//	Se houver erro na validação
+			    return "forward:adicionar_pedido";								//	Volte para a página de adição
+			} else {
+				dao.create(pedido, carrinho, cliente);							//	Ação no banco
+				return "redirect:pedido";										//	Retorna para o método Read
+			}
 		}
-		
+		return "403";
 	}
 
 	/*
@@ -84,9 +103,12 @@ public class PedidoController {
 	 */
 	
 	@RequestMapping("Admin/pedido")
-	public String Read(Model model) {
-		model.addAttribute("pedidos", dao.read());							//	Consulta o Banco e coloca na variável da página
-		return "admin/Pedido/read";											//	Retorna para á página JSP
+	public String read(HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.LER, entidade)){			//	Consulta Permissão
+			model.addAttribute("pedidos", dao.read());						//	Consulta o Banco e coloca na variável da página
+			return "admin/Pedido/read";										//	Retorna para á página JSP
+		}
+		return "403";
 	}
 
 	/*
@@ -103,14 +125,17 @@ public class PedidoController {
 	 */
 
 	@RequestMapping("Admin/UpdatePedido")
-	public String update(@Valid Pedido pedido, BindingResult result) {
-
-		if(result.hasErrors()) {											//	Se houver erro na validação
-		    return "forward:pedido";										//	Volte
-		} else {
-			dao.update(pedido);												//	Ação no banco
-			return "redirect:pedido";										//	Retorna para o método Read
+	public String update(HttpSession session, @Valid Pedido pedido, BindingResult result) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){		//	Consulta Permissão
+			if(result.hasErrors()) {										//	Se houver erro na validação
+			    return "forward:pedido";									//	Volte
+			} else {
+				dao.update(pedido);													//	Ação no banco
+				relatorio.gerarRelatorio(session, EnumMetodo.ATUALIZAR, entidade);	//	Gera Relatório e armazena no banco
+				return "redirect:pedido";											//	Retorna para o método Read
+			}						
 		}
+		return "403";
 	}
 
 	/*
@@ -127,9 +152,13 @@ public class PedidoController {
 	 */
 	
 	@RequestMapping("Admin/DeletePedido")
-	public void delete(Long id) {
-	  dao.delete(id);														//	Ação no banco
-	}
+	public void delete(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.EXCLUIR, entidade)){			//	Consulta a permissão
+			dao.delete(id);														//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.EXCLUIR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
+		}
+	} 
 	
 	/*
 	 * -------------------------
@@ -145,9 +174,14 @@ public class PedidoController {
 	 */
 	
 	@RequestMapping("Admin/RestorePedido")
-	public void restore(Long id) {
-		  dao.restore(id);													//	Ação no banco
+	public void restore(HttpSession session, Long id, HttpServletResponse response) {
+		if(Permissoes.checar(session, EnumMetodo.RESTAURAR, entidade)){			//	Consulta a permissão
+			dao.restore(id);													//	Ação no banco
+			relatorio.gerarRelatorio(session, EnumMetodo.RESTAURAR, entidade);	//	Gera Relatório e armazena no banco
+			response.setStatus(200);											//	Indica para a requisição AJAX que tudo ocorreu bem
+		}
 	}
+
 
 
 	/*
@@ -164,11 +198,18 @@ public class PedidoController {
 	 */
 	
 	@RequestMapping("Admin/FindPedido")
-	public String find(Model model, Long id) {
-			model.addAttribute("pedido", dao.findOne(id));					//	Consulta o Banco e coloca na variável da página
-			model.addAttribute("produtos", dao_prod.read());				//	Consulta o Banco e coloca na variável da página
-			model.addAttribute("clientes", dao_cli.read());					//	Consulta o Banco e coloca na variável da página
-			return "admin/Pedido/edt";										//	Retorna para a página JSP edt
+	public String Find(Long id, HttpSession session, Model model) {
+		if(Permissoes.checar(session, EnumMetodo.ATUALIZAR, entidade)){			//	Consulta a permissão
+			model.addAttribute("pedido", dao.findOne(id));							//	Consulta o Banco e coloca na variável da página
+			model.addAttribute("produtos", dao_prod.read());					//	Consulta o Banco e coloca na variável da página
+			model.addAttribute("clientes", dao_cli.read());						//	Consulta o Banco e coloca na variável da página
+			return "admin/Pedido/edt";											//	Retorna para a página JSP edt
+		}
+		return"403";
 	}
-	
 }
+
+
+		  
+		
+
